requirements.txt
# requirements.txt
httpx>=0.27.0
fastapi
uvicorn
python-multipart
python-dotenv
pandas
psycopg2-binary
matplotlib
SQLAlchemy>=2.0
asyncpg
alembic
greenlet
